refactor: CategoryTag, CategoryTags,  TextWithIcon, 공통 컴포넌트로 이동
