+++ frontend/src/common/components/TextWithIcon/TextWithIcon.tsx
